VOTE-1947 Reorganize repo to follow react best practices
