+++ src/App.jsx
@@ -1,12 +1,12 @@
 import { useState, useEffect } from 'react'
-import StateSelection from './components/StateSelection';
-import Eligibility from './components/Eligibility';
-import RegistrationOptions from './components/RegistrationOptions';
-import PathSelection from './components/PathSelection';
-import MultiStepForm from './components/MultiStepForm';
-import {fetchData, sanitizeDOM} from './components/HelperFunctions/JsonHelper.jsx';
+import StateSelection from './views/StateSelection.jsx';
+import Eligibility from './views/Eligibility.jsx';
+import RegistrationOptions from './views/RegistrationOptions.jsx';
+import PathSelection from './views/PathSelection.jsx';
+import MultiStepForm from './views/MultiStepForm.jsx';
+import {fetchData, sanitizeDOM} from './utils/JsonHelper.jsx';
 import { HelmetProvider } from "react-helmet-async";
-import {getFieldValue} from "./components/HelperFunctions/fieldParser";
+import {getFieldValue} from "./utils/fieldParser.jsx";
 
 function App() {
