+++ src/components/RegType/ByMail.jsx
@@ -1,7 +1,7 @@
 import { Icon } from '@trussworks/react-uswds';
-import NextButton from '../NextButton';
+import NextButton from '../Buttons/NextButton';
 import {renderToStaticMarkup} from "react-dom/server";
-import {sanitizeDOM} from "../HelperFunctions/JsonHelper";
+import {sanitizeDOM} from "../../utils/JsonHelper";
 
 function ByMail(props) {
     const content = props.content;
