+++ src/components/RegType/NotNeeded.jsx
@@ -1,6 +1,6 @@
 import { Link, Icon } from '@trussworks/react-uswds';
 import {renderToStaticMarkup} from "react-dom/server";
-import {sanitizeDOM} from "../HelperFunctions/JsonHelper";
+import {sanitizeDOM} from "../../utils/JsonHelper";
 
 function NotNeeded(props) {
     const content = props.content;
